Move all volume reconstruction code into separate files
There is no code change, just moving code around and preparing for the
subsequent commit.
